+++ CHANGELOG.md
@@ -17,6 +17,8 @@ is roughly based on [Keep a Changelog], and this project adheres to
 
 ### Fixes
 
+- Fixes bug causing alternative or uppercase module names to fail to load the corresponding SSWSort module
+
 ## [2.0.0] - 2026-07-08
 
 ### Changed