+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file. The format
 is roughly based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [2.0.2] - TBD
+
+### Added
+
+- `zoe_genome_aln` is now a field of `GenomeAndProductStates` to provide access to the genome alignment and CIGAR string
+
 ## [2.0.1] - 2026-07-10
 
 ### Added
@@ -200,6 +206,7 @@ is roughly based on [Keep a Changelog], and this project adheres to
 
 <!-- Versions -->
 
+[2.0.2]: https://github.com/CDCgov/dais-ribosome/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/CDCgov/dais-ribosome/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/CDCgov/dais-ribosome/compare/v1.7.1...v2.0.0
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/